Write custom resource manager for handling http requests
The custom resource manager always returns an instance of the custom resource class pointing to the current output file.
The custom resource overrides the 'getName' function to always return the path that was requested.

For example if you go to '123.45.67.89:8080/cool/stuff/myServerPack.zip', it will download the latest pack and name it 'myServerPack.zip'